backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-JWT-Extended==4.6.0
python-dotenv==1.0.0
supabase==1.2.0
Werkzeug==2.3.7
httpx==0.24.1
PyMuPDF
openai==1.3.0
requests==2.31.0
